tty/serial: don't suspend xen serial port
Remove PNP_DISABLE capability for serial port on 0x3f8 as Xen uses it
and will disable it when going to S3.

Disables the functionality to evaluate ACPI _DIS method on serial 0x3f8
which is used by Xen logging. Fixes S3 for machines with PNP serial
port. Xen is supposed to disable the serial port as it uses it.