xen-netfront: close fe if be disappears
Treat a vanishing backend as a request to close the front-end.

Usual scenario would be dom0 releaving netback responsibility to NDVM
ending up to rmmod its backend module. Xenstore node would disapear and
Xenbus switch to unknown state. We might as well consider it, on the
frontend side (e.g, in UIVM) to be a close request, so the guest will be
able to reconnect with the new backend in the service VM.