xen: paravirtualize io permission bitmap
Origin:
http://old-list-archives.xen.org/archives/html/xen-devel/2009-05/msg01139.html

A PV Xen guest kernel has no TSS of its own, so the IO permission bitmap
must be paravirtualized. This patch adds set_io_bitmap as a paravirt op,
and defines a native version which updates the tss, and a Xen version
which uses a hypercall.

This is much easier now that 32 and 64-bit use the same code to manage
the IO bitmap.