tpm_tis: force ioremap of the io region
This region is only reported in one of the SSDT, not the e820. On a
system using a 32bit kernel and enough memory, the kernel might end up
making this region a RAM buffer, disabling the TPM.

devm_ioremap() was used (prior 4.6?), avoiding the
devm_request_mem_region() that currently fails under the mentioned
circumstances. Fallback to that until a better fix is found.