drm/i915: add gem_foreign ioctl.
DRM_IOCTL_I915_GEM_FOREIGN will return a GEM object created with pages
from a foreign guest. That can then be used by the current domain DRM
engine to be scanned. The goal is to let a guest soft-render in a
surface while another guest is driving the GPU (dom0).

This is limited to i915 devices.