squash in 99a25c12dfae
linux: make usbback run in an HVM backend

usbback is limited to only loading in PV domains. If that check is
removed, a runtime error is hit in set_phys_to_machine() since that call
is inapproriate for an HVM domain.

The grant mapping and unmapping code open coded to use the raw hypercall
instead of the proper gnttab api. Switching to the gnttab API properly
skips set_phys_to_machine for HVM domains. To use that API, an
contiguous array of struct page *'s is needed instead of the scattered
ones in pending_segment_t. Also use of pages obtained by alloc_page
fails in set_foreign_p2m_mapping with "page must be ballooned". Switch
to gnttab_alloc_pages to get the proper pages.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>

---
Maybe pending_segment_t should lose its struct page and the array be
embedded in pending_req_t. That would require more changes, so it is not
done here.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>